Time: 12 ms (19.77%), Space: 20.1 MB (7.90%) - LeetHub

+++ 0010-regular-expression-matching/0010-regular-expression-matching.cpp
@@ -1,13 +1,8 @@
 class Solution {
 public:
 
-    bool solve(int index1,int index2,string s,string p,vector<vector<int>>&dp)
+    bool solve(long index1,long index2,string s,string p,vector<vector<int>>&dp)
     {
-        //Hello 
-        //This
-        //is
-        //a
-        //code
         if(index1<0 && index2<0)
         {
             return true;
